requirements.txt
# Core
PyQt6==6.6.0
pyserial==3.5

# Dev-Tools (optional)
pytest==7.4.0
black==23.9.1
flake8==6.0.0